Fix MockGenerator for TerminatableMixin.ensureNotTerminatedSync
