+++ pubspec.yaml
@@ -7,7 +7,7 @@ environment:
   sdk: '>=2.12.0 <3.0.0'
 
 dev_dependencies:
-  build_runner: ^1.10.0
-  mockito: ^5.0.0-nullsafety
+  build_runner: '>=1.10.0 <3.0.0'
+  mockito: ^5.0.9
   test: ^1.16.0
   pedantic: ^1.10.0
